Changed the report table.
Changed the format to a list of lists of length 3 and added empty report lists for the revenue and budget reports.

The reports are added to the one end of game report table.

Random tweaking with colours and doing special rows in the table by taking them out of for loops to change the colours.